+++ client/src/components/campaign/Party/PlayerList.jsx
@@ -1,17 +1,14 @@
 import React from "react";
 
 const PlayerList = () => {
-
   return (
     <section>
       <div className="bg-primary rounded-xl text-textcolor m-2">
         <h1 className="text-2xl m-4 p-5">Party Details</h1>
       </div>
-      <div>
-        {/* IMAGES HERE - BASE , CARDS - STRETCH */}
-      </div>
+      <div>{/* IMAGES HERE - BASE , CARDS - STRETCH */}</div>
     </section>
-  )
-}
+  );
+};
 
 export default PlayerList;
